Added .ai and .svg design files
I've created 2 new subfolders to make it easier for designers, developers, and Nostr advocates to easily find the filetypes they are looking for.

We could also do this for the .png files as well. Organizing the directory will make searching and finding the exact design much easier. PS: Love the Figma board.